backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Document processing
PyMuPDF==1.23.8
chardet==5.2.0

# Embeddings and vector store
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.26.3

# LLM integration
requests==2.31.0

# Utilities
pydantic==2.5.3
python-dotenv==1.0.0
